Force ELF load-time reloc generation in executables
Non-PIE executables on most systems are generally built to run at a
fixed address. Since this isn't the case on z/OS, we have to rely on
load-time relocation of one sort or another. For static executables,
we already generate Program Object runtime relocations as necessary,
however for dynamically linked position-dependent executables we need
to take a different approach.

Executables must be able to run when invoked as an explicit argument to
the dynamic linker. The dynamic linker is not required to go through the
normal z/OS program loading facilities, so it cannot rely on the Program
Object load-time relocations alone.

The solution is to generate normal s390x ELF load-time relocations
alongside the Program Object relocations. Unfortunately, the
no-interference hooking approach that we use to generate the Program
Object relocations will not work for ELF relocations, so we need add in
some slightly more explicit and invasive code into elf64-s390.c (though
nothing that actually affects the functionality of the elf64-s390
emulation itself).

INCU-209