Always emit dyn relocs for resolved GOT entries
Add more macros to hook into the static resolution of GOT entries. They
ensure that whenever the backend would normally try to fill in the
in the GOT entry with the final address of the target symbol, it instead
generates both PO and ELF load-time relocs. We are also careful to
reserve space in the reloc sections ahead of time.

INCU-229