Allow arbitrary size Program Object headers
Modify the Program Object header generation so that the headers
only take up as much space as required, and do not have a maximum size.
Previously, as an implemenation convenience, we would reserve a large
area for the headers and failed if that was not enough, which was both
unreliable for very large executables and suboptimal for small
executables, as the reserved area was a significant waste of space and
memory. This change reduces the filesize and runtime memory image of
all executables by up to 2MB, with savings being more dramatic the
smaller the executable.

INCU-331