Fix relocs not being generated for some cases
When the PO code causes ELF dynamic relocs to be generated when they
would not be otherwise, sometimes the symbol would not be relocated
normally, leading to no PO relocation being generated. This change
should make sure that whenever we force ELF dynamic relocs to be
generated, the relevant relocs should still be relocated normally.

INCU-433